购物车增加和删除

+++ src/main/java/com/furuida/model/ShoppingCart.java
@@ -5,7 +5,7 @@ public class ShoppingCart {
 
     private String userId;
 
-    private Byte goodsId;
+    private Integer goodsId;
 
     private Integer num;
 
@@ -29,11 +29,11 @@ public class ShoppingCart {
         this.userId = userId == null ? null : userId.trim();
     }
 
-    public Byte getGoodsId() {
+    public Integer getGoodsId() {
         return goodsId;
     }
 
-    public void setGoodsId(Byte goodsId) {
+    public void setGoodsId(Integer goodsId) {
         this.goodsId = goodsId;
     }
